remove XML formats from minimal
